Split out Instrumentation Gateway calling code from HttpCodec
