+++ dd-java-agent/agent-bootstrap/src/main/java/datadog/trace/bootstrap/instrumentation/decorator/HttpServerDecorator.java
@@ -6,7 +6,10 @@ import static datadog.trace.bootstrap.instrumentation.decorator.RouteHandlerDeco
 
 import datadog.trace.api.Config;
 import datadog.trace.api.DDTags;
+import datadog.trace.api.Function;
 import datadog.trace.api.function.BiFunction;
+import datadog.trace.api.function.Supplier;
+import datadog.trace.api.function.TriConsumer;
 import datadog.trace.api.function.TriFunction;
 import datadog.trace.api.gateway.CallbackProvider;
 import datadog.trace.api.gateway.Events;
@@ -17,6 +20,7 @@ import datadog.trace.bootstrap.instrumentation.api.AgentPropagation;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import datadog.trace.bootstrap.instrumentation.api.AgentTracer;
 import datadog.trace.bootstrap.instrumentation.api.InternalSpanTypes;
+import datadog.trace.bootstrap.instrumentation.api.TagContext;
 import datadog.trace.bootstrap.instrumentation.api.Tags;
 import datadog.trace.bootstrap.instrumentation.api.URIDataAdapter;
 import datadog.trace.bootstrap.instrumentation.api.URIUtils;
@@ -70,19 +74,24 @@ public abstract class HttpServerDecorator<REQUEST, CONNECTION, RESPONSE, CARRIER
     return Config.get().isTraceAnalyticsEnabled();
   }
 
+  // Extract this to allow for easier testing
+  protected AgentTracer.TracerAPI tracer() {
+    return AgentTracer.get();
+  }
+
   public AgentSpan.Context.Extracted extract(CARRIER carrier) {
     AgentPropagation.ContextVisitor<CARRIER> getter = getter();
     if (null == carrier || null == getter) {
       return null;
     }
-    return AgentTracer.propagate().extract(carrier, getter);
+    return tracer().propagate().extract(carrier, getter);
   }
 
   public AgentSpan startSpan(CARRIER carrier, AgentSpan.Context.Extracted context) {
-    // TODO move AppSec header checks here after the span has been created, and also create the
-    //  RequestContext after the span has been created, or maybe even move it after the URL is
-    //  available?
-    return AgentTracer.startSpan(spanName(), context).setMeasured(true);
+    AgentSpan span =
+        tracer().startSpan(spanName(), callIGCallbackStart(context), true).setMeasured(true);
+    callIGCallbackHeaders(span, carrier);
+    return span;
   }
 
   public AgentSpan onRequest(
@@ -195,10 +204,51 @@ public abstract class HttpServerDecorator<REQUEST, CONNECTION, RESPONSE, CARRIER
   //    return super.onError(span, throwable);
   //  }
 
-  private static void callIGCallbackURI(
+  private AgentSpan.Context.Extracted callIGCallbackStart(AgentSpan.Context.Extracted context) {
+    CallbackProvider cbp = tracer().instrumentationGateway();
+    if (null != cbp) {
+      Supplier<Flow<RequestContext>> startedCB = cbp.getCallback(Events.REQUEST_STARTED);
+      if (null != startedCB) {
+        RequestContext requestContext = startedCB.get().getResult();
+        if (null != requestContext) {
+          TagContext tagContext = null;
+          if (context == null) {
+            tagContext = TagContext.empty();
+          }
+          if (context instanceof TagContext) {
+            tagContext = (TagContext) context;
+          }
+          if (null != tagContext) {
+            context = tagContext.withRequestContext(requestContext);
+          }
+        }
+      }
+    }
+    return context;
+  }
+
+  private void callIGCallbackHeaders(AgentSpan span, CARRIER carrier) {
+    CallbackProvider cbp = tracer().instrumentationGateway();
+    RequestContext requestContext = span.getRequestContext();
+    AgentPropagation.ContextVisitor<CARRIER> getter = getter();
+    if (requestContext == null || cbp == null || getter == null) {
+      return;
+    }
+    IGKeyClassifier igKeyClassifier =
+        IGKeyClassifier.create(
+            requestContext,
+            cbp.getCallback(Events.REQUEST_HEADER),
+            cbp.getCallback(Events.REQUEST_HEADER_DONE));
+    if (null != igKeyClassifier) {
+      getter.forEachKey(carrier, igKeyClassifier);
+      igKeyClassifier.done();
+    }
+  }
+
+  private void callIGCallbackURI(
       @Nonnull final AgentSpan span, @Nonnull final URIDataAdapter url, final String method) {
     // TODO:appsec there must be some better way to do this?
-    CallbackProvider cbp = AgentTracer.get().instrumentationGateway();
+    CallbackProvider cbp = tracer().instrumentationGateway();
     RequestContext requestContext = span.getRequestContext();
     if (requestContext == null || cbp == null) {
       return;
@@ -217,11 +267,11 @@ public abstract class HttpServerDecorator<REQUEST, CONNECTION, RESPONSE, CARRIER
     return super.beforeFinish(span);
   }
 
-  private static void onRequestEndForInstrumentationGateway(@Nonnull final AgentSpan span) {
+  private void onRequestEndForInstrumentationGateway(@Nonnull final AgentSpan span) {
     if (span.getLocalRootSpan() != span) {
       return;
     }
-    CallbackProvider cbp = AgentTracer.get().instrumentationGateway();
+    CallbackProvider cbp = tracer().instrumentationGateway();
     RequestContext requestContext = span.getRequestContext();
     if (cbp != null && requestContext != null) {
       BiFunction<RequestContext, IGSpanInfo, Flow<Void>> callback =
@@ -232,9 +282,9 @@ public abstract class HttpServerDecorator<REQUEST, CONNECTION, RESPONSE, CARRIER
     }
   }
 
-  private static Flow<Void> callIGCallbackSocketAddress(
+  private Flow<Void> callIGCallbackSocketAddress(
       @Nonnull final AgentSpan span, @Nonnull final String ip, final int port) {
-    CallbackProvider cbp = AgentTracer.get().instrumentationGateway();
+    CallbackProvider cbp = tracer().instrumentationGateway();
     if (cbp == null) {
       return Flow.ResultFlow.empty();
     }
@@ -247,5 +297,44 @@ public abstract class HttpServerDecorator<REQUEST, CONNECTION, RESPONSE, CARRIER
       }
     }
     return Flow.ResultFlow.empty();
+  }
+
+  /** This passes the headers through to the InstrumentationGateway */
+  private static final class IGKeyClassifier implements AgentPropagation.KeyClassifier {
+
+    private static IGKeyClassifier create(
+        RequestContext requestContext,
+        TriConsumer<RequestContext, String, String> headerCallback,
+        Function<RequestContext, Flow<Void>> doneCallback) {
+      if (null == requestContext || null == headerCallback) {
+        return null;
+      }
+      return new IGKeyClassifier(requestContext, headerCallback, doneCallback);
+    }
+
+    private final RequestContext requestContext;
+    private final TriConsumer<RequestContext, String, String> headerCallback;
+    private final Function<RequestContext, Flow<Void>> doneCallback;
+
+    private IGKeyClassifier(
+        RequestContext requestContext,
+        TriConsumer<RequestContext, String, String> headerCallback,
+        Function<RequestContext, Flow<Void>> doneCallback) {
+      this.requestContext = requestContext;
+      this.headerCallback = headerCallback;
+      this.doneCallback = doneCallback;
+    }
+
+    @Override
+    public boolean accept(String key, String value) {
+      headerCallback.accept(requestContext, key, value);
+      return true;
+    }
+
+    public void done() {
+      if (null != doneCallback) {
+        doneCallback.apply(requestContext);
+      }
+    }
   }
 }
