+++ dd-trace-core/src/main/java/datadog/trace/core/DDSpanContext.java
@@ -353,7 +353,6 @@ public class DDSpanContext implements AgentSpan.Context {
     return trace;
   }
 
-  @Override
   public RequestContext getRequestContext() {
     return requestContext;
   }
